Reserva diferida de páginas

+++ xv6/trap.c
@@ -8,6 +8,8 @@
 #include "traps.h"
 #include "spinlock.h"
 
+
+extern int mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm);
 // Interrupt descriptor table (shared by all CPUs).
 struct gatedesc idt[256];
 extern uint vectors[];  // in vectors.S: array of 256 entry pointers
@@ -86,13 +88,30 @@ trap(struct trapframe *tf)
               tf->trapno, cpuid(), tf->eip, rcr2());
       panic("trap");
     }
+  if(tf->trapno == T_PGFLT){ //page fault treatment
+    if (rcr2() < myproc()->sz){ // check that we are on range
+      char *mem;//page
+      if((mem = kalloc()) == 0){ // error on memory == death
+        myproc()->killed = 1;
+        return;
+      }
+      memset(mem, 0, PGSIZE); // put every bit on the page to 0
+      if(mappages(myproc()->pgdir,(void *) PGROUNDDOWN(rcr2()) , PGSIZE, V2P(mem) , PTE_W|PTE_U) < 0){
+        cprintf("trap: out of memory (2)\n");
+        kfree(mem);
+        myproc()->killed = 1;
+        return;
+      }
+    }else{
     // In user space, assume process misbehaved.
     cprintf("pid %d %s: trap %d err %d on cpu %d "
             "eip 0x%x addr 0x%x--kill proc\n",
             myproc()->pid, myproc()->name, tf->trapno,
             tf->err, cpuid(), tf->eip, rcr2());
-    myproc()->killed = 1;
+        myproc()->killed = 1;
+    }
   }
+}
 
   // Force process exit if it has been killed and is in user space.
   // (If it is still executing in the kernel, let it keep running
@@ -110,3 +129,52 @@ trap(struct trapframe *tf)
   if(myproc() && myproc()->killed && (tf->cs&3) == DPL_USER)
     exit();
 }
+
+
+/*if(tf->trapno == T_PGFLT){ //page fault treatment
+      if (rcr2()< myproc()->sz){ // check that we are on range
+        char *mem = kalloc();
+        if(mem == 0){ // error on memory == death
+          myproc()->killed = 1;
+          return;
+        }
+        memset(mem, 0, PGSIZE); // put every bit on the page to 0
+        if(mappages(myproc()->pgdir, PGROUNDUP(rcr2()), PGSIZE, (uint)mem, PTE_W | PTE_U) < 0){ // error on memory 2 == death
+          cprintf("trap: out of memory (2)\n");
+          kfree(mem);
+          myproc()->killed = 1;
+          return;
+        }
+      }else
+      {
+        cprintf("out of range page at virtual addres 0x%s\n", rcr2);
+        myproc()->killed = 1;
+        return;
+      }
+    }else */
+
+
+    /*if ( tf->trapno == T_PGFLT )
+    {
+      if ( rcr2 () >= myproc () -> sz )
+      {
+        cprintf ( " La pagina que ha fallado esta fuera del rango \n " ) ;
+        myproc () -> killed = 1;
+        break ;
+      }
+      char * marco = kalloc () ;
+      if ( marco == 0) {
+        cprintf ( " Fallo de asignacion de memoria trap \n " ) ;
+        myproc () -> killed = 1;
+        return ;
+      }
+      memset ( marco , 0 , PGSIZE ) ;
+      if ( mappages ( myproc () -> pgdir ,( void *) PGROUNDDOWN ( rcr2
+      () ) , PGSIZE , V2P ( marco ) , PTE_W | PTE_U ) < 0)
+      { 
+        cprintf ( " Fallo en la asignacion con mappages \n " ) ;
+        kfree ( marco ) ;
+        myproc () -> killed = 1;
+        return ;
+      }
+   }else { */
